# cuda4dnn: do not create temporary unused handles