Layers that require cuDNN or cuBLAS capabilities have handles as member object. A useless handle is created when they are default constructed. They are going to be replaced by the actual handle later. Handle creation is costly. Hence, all the temporary handles waste resources.

Each `cv::dnn::Net` object now creates exactly one stream, one cuDNN handle and one cuBLAS handle. 

This PR makes default constructed handle objects not create any handle. A special constructor must be used to create a handle.

Model                      | init time (master) | init time (with this PR)
-------------------------- | ---------------------- | -----------------------------
ResNet 50               | 260ms                 | 170ms                        
Enet Cityscapes      | 170ms                 | 30ms
YOLOv4                  | 630ms                 | 440ms
Inception v2 Mask RCNN            | 520ms                 | 340ms

**Others:**
- drop support for default streams (for better error detection)
- do  not allow moving empty handles into other handles (for better error detection)
- fix handle leak (no bugs yet since the bugged move constructor isn't being used)
- avoid creating shared pointers for empty UniqueHandle objects

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under OpenCV (BSD) License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [ ] There is reference to original bug report and related work
- [ ] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [ ] The feature is well documented and sample code can be built with the project CMake

```
force_builders=Custom
buildworker:Custom=linux-4
build_image:Custom=ubuntu-cuda:18.04
buildworker:iOS=macosx-2
```