@YashasSamaga 2020-05-22
I had planned on fixing `getPerfProfile` for the CUDA backend. Currently, the layerwise timings returned by `getPerfProfile` is the time taken for dumping the layer operation into the CUDA stream. It does not tell how much time the operation actually took.

The only way to allow proper `getPerfProfile` result is to place events before and after every layer in the CUDA stream. But this comes at a cost which appears to be on average ~1% (need to redo calculations on idle system) on the inference time. I am not sure what to do.

Many samples and tests use `getPerfProfile`. They all lead to wrong results when the CUDA backend is used.